Updated VoxelParse.py
Edited .py script to loop over .pdb files in a subdirectory named 'pdb', of the current working directory.
Moved settings section up.
Line 71: @Camille, this part is buggy for me. 'structure.get_residues():' worked for me
Edited also to create a 'pickle' directory and save all pickles (same naming as the pdb file).
I don't have the atom_density script so I have not been able to run this whole script with success, but the specific sections with my changes seemed to work out in the notebook.

Pending: Will make corresponding changes to Code 2 so it takes in the pickles in a loop as well.